fix: controller flow
