+++ app/dto/ScheduleTTSRequestDto.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 
 
 class ScheduleTTSRequestDto(BaseModel):
-    voice_id: int
+    voice_id: str
     alias: str
     schedule_id: List[int]
     schedule_text: List[str]
